requirements.txt
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.3
gevent==24.11.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
logzero==1.7.0
numpy==2.1.3
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
smartapi-python==1.5.1
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket==0.2.1
zope.event==5.0
zope.interface==7.1.1
